+++ jobs/Frontend/movies/.env.sample
@@ -1,4 +1,4 @@
 ## Sample configuration
 
 ## TheMovieDb API key
-# VITE_FMDB_API_KEY=ABC
+# VITE_TMDB_API_KEY=ABC
