+++ jobs/Frontend/movies/src/config.ts
@@ -1,1 +1,1 @@
-export const FMDB_API_KEY = import.meta.env.VITE_FMDB_API_KEY || '';
+export const TMDB_API_KEY = import.meta.env.VITE_TMDB_API_KEY || '';
