+++ jobs/Frontend/movies/src/vite-env.d.ts
@@ -1,2 +1,2 @@
 /// <reference types="vite/client" />
-declare const FMDB_API_KEY: string;
+declare const TMDB_API_KEY: string;
